refactoring universal table
